Fix SelectedAdministrativeCriteria.certification_period
The API particulier computed the dateFin with dateDebut + 3 months.
That’s very misleading, as a person who has been a beneficiary since
2009-06-01 would have:
```json
{
    "statut": "beneficiaire",
    "dateDebut": "2009-06-01",
    "dateFin": "2009-09-01",
}
```

However, the `statut` means the person is beneficiary **today**.

The API recently evolved to always set `dateFin` to `null` to avoid
suggesting the end date is known and in the past.

Fix about 11,000 incorrect records, by using `certified_at`: if the API
response for `statut` was "beneficiaire", that person was beneficiary on
the day the API was queried.

The example from above comes from the production environment and is an
example of bad data.