gps: don't use `apply` to bounce on the benefiary account creation
Creating a new beneficiary does not need to pass by `apply` anymore.
Now using `job_seekers_views:get_or_create_start` with `tunnel="gps"`.

In this commit, the logic and permission checking in `apply` is kept in
order to avoid errors at deploy.

Now the "Annuler" button sends back to gps:my_group and not to a 404.