job_seekers_views: raise a 403 error when using a wrong user kind
In job_seekers_views, GetOrCreate views come in two different flavors:
- CheckNIRForJobSeeker, only for job seekers
- CheckNIRForSender, SearchByEmailForSender etc, for senders/hire

It is virtually impossible to be on the wrong flavor by mistake, because
of the session system: we check the user kind and create a session that
can be used on that flavor only.

However we keep a safety rail: a 403 error when a user kind does not
match the current flavor.