Use JobSeekerProfileFieldMixin in JobSeekerPersonalDataForm
Simplify the inheritance for JobSeekerNIRUpdateMixin, as all forms are
now based on the User model, and consolidate implementation of user
profile fields on JobSeekerProfileFieldMixin.

Because fields are now added to JobSeekerPersonalDataForm dynamically by
the JobSeekerProfileFieldMixin, `base_fields` are empty. Use the
birthdate field on the form instance to perform the validation. That’s
also a bit safer, as we don’t risk making changes to the reused
base_fields for other requests.