Prevent editing job seeker with a certified eligibility criteria
The job seeker identity is used to certify eligibility against the API
Particuliers. When eligibility is certified, the hire will be removed
from siae evaluations.

In order to prevent employers to reuse an identity for multiple job
seekers, make sure the job seeker personal data is locked when their
eligibility has been confirmed.

To better show the dependency between read-only form fields and the API
Particuliers, factor out the required fields for the API into constants
that can be reused across the codebase.

TODO:
- SSO (and tests)