Make JobSeekerProfile birth_{country,place} more realistic
Ensure the birth_place start_date and end_date attribute are around the
birthdate, so that data is realistic.

Updated the factory to use actual objects from the ASP and not a fakes,
as the application doesn’t manage Commune and Country objects.

Remove the with_birth_place trait, as it would setup an incomplete
object (birth country isn’t France). It was only used once and was
merged into the born_in_france trait.