admin: Make memberships RO when there are too many
For organization with more than 250 members, we go over the
DATA_UPLOAD_MAX_NUMBER_FIELDS limit of 1000.

Make the memberships readonly when they have too many (let supportix or
another member of the organization add the user)
We have less than 10 organizations with more than 200 members

No need to make this change for companies and institutions
since we have respectively at most 42 and 23 memberships
as of today.