users: add a method to ItouUserManager to get the list of job seeker ids
This is used to get a list of a user's job seeker ids (ie. job seekers
created by the user, or by the current organization, or for whom the
user has applied).

This logic was located in a view in job_seekers_views but is better in
the ItouUserManager, in order to be used as
`User.objects.linked_job_seeker_ids(user, organization)`