Split new_users_to_brevo processing
Mainly, this refactoring aims to facilitate integration of job seekers
in the command.

Avoids keeping all results in memory, when a batch is identified and
sent, all references to the batch are dropped and memory will be freed.

The ordering of logs changed slightly due to the reordering of
operations.