login: Replace admin login view
Using the perms middleware.
We can now go to a specific admin page (eg /admin/users/user/PK) and be
redirected to a login page that will redirect us back to the correct
page.
Before this commit, when accessing such page, we would have the generic
login page that does not have a "staff login" link