Bump boto3
Cellar is not at feature parity with AWS S3. AWS S3 can validate
integrity of chunks transferred with a checksum, and that has been
enabled by default in boto3. Cellar does not support that feature, and
attempting to upload a document errors out with a MissingContentLength.
https://github.com/boto/boto3/issues/4392

Disable the integrity checks for Cellar.

Although the Pilotage S3 bucket is on Scaleway, which may use an S3
implementation that is more in line with AWS S3, it runs its tests
against les-emplois infrastructure, which is on Cellar, so it has been
configured not to run the integrity checks as well.

The signature version that was specified in s3_client() is actually the
default, best omit it.
https://boto3.amazonaws.com/v1/documentation/api/latest/guide/configuration.html#:~:text=signature_version

The xfail test will fail when Cellar supports the integrity checks.