Run CI against Cellar
https://github.com/gip-inclusion/itou-communaute-django/issues/897 also
hit les-emplois.

Basically, a new version of Boto3 enables features that are supported by
AWS S3, MinIO, but not Cellar. The problem goes undetected until
production, because tests are run against MinIO and may cause an outage.

In order to prevent such issues in the future, run the CI against a
Cellar bucket. The local environment keeps using MinIO so that tests run
offline, and to avoid setting up credentials locally.