Verify session kind in SessionNamespace
Multiple forms on the site are quite long, and were split over multiple
pages, storing the intermediary data in the session. Data is only
written to the database at the end of the editing process.

In order to work well with tabs, the sessions are namespaced (usually
stored under a UUID at the root of the session), and the user passes the
namespace through the URL. Which means the user can read the UUID from
the URL, and try reusing a session from an editing process in other
views, allowing users to somewhat manipulate their session.

Although most processes are expected to fail when given data from
another editing process, it remains a security risk as developers cannot
reasonably keep track of the intersection of session data for all
editing processes on the site.

To protect against such issues, some views (subclasses of WizardView)
store a session_kind, and verify it when they load the user session.

Generalize the verification to all `SessionNamespace`, to force all
users of `SessionNamespace` to follow that pattern.