companies: transfer memberships even if a membership exists in to_company
In the admin, when transfering a company to another one, do transfer the
memberships even if a membership with the same user exists.
The rule for *overwriting* `is_active` and `is_admin` is `any` (ie for a
given user, set is_admin=True if any of the memberships is admin, and
the same goes for is_active).