siae_evaluations: Automatically select certified administrative criteria
When a criterion is certified, it does not need verification by the
DDETS. When the evaluation campaign starts, all certified criteria are
added as an accepted evaluated administrative criteria. If these
criteria are enough for all selected job applications to be accepted,
the evaluated SIAE is immediately set to accepted, and the SIAE isn’t
even notified of the control: they have nothing to do.