Eligibility: Fix administrative criteria is_considered_certified
Originally, the API particulier provided a start date and an end date
(computed as the current date + 3 months). It was updated to always
return None for the end date field, to better indicate that the end date
is not known without breaking existing integrations.

The certification period was updated to follow the API changes with
9d44cd13f602550e2aae38afc6a3487897abc704, where the end date is now
computed to 92 days after the certification date, as the API only
answers the question “is that person a beneficiary today?”. The 92 days
come from our purpose for certification: it’s the 3-month limit to
justify access to the RSA.

However, we did not go back to `is_considered_certified`, which also
grants the 92 days grace period, resulting in an 184 days grace period.