Verify certified badge isn’t shown after diagnosis expiration
The computation of certification expiry was incorrect, a criteria was
considered valid much longer than intended. This bug hasn’t surfaced to
employers though, because the eligibility diagnosis validity period has
been adjusted to match the certification period, so the diagnosis
expired even though the certification was still valid.

This commit adds a test to confirm employers did not see the certified
badge.

Refs https://github.com/gip-inclusion/les-emplois/pull/6126