siae_evaluations: Merge job_seeker_infos into a template tag
Previously, job_seeker_infos would use two templates, depending on the
viewer. Each template was pretty complex, relying on if precedence to
display the correct state, with multiple nested branches.

Instead, port the code to python and organize it to match the evaluation
stages. At each stage, a dict maps the state of the SIAE with the
displayed state, hopefully making the logic easier to track and verify.