siae_evaluations: Remove PENDING_AFTER_REVIEW_BADGE
Once upon a time, siae_evaluations didn’t follow a calendar, the whole
evaluation process could be completed in an hour. It has since been
updated to follow 6 weeks periods for employers to upload proofs, and at
least 2 weeks periods for DDETS to review them.

In that distant past, it was useful to distinguish the TODO state in the
amicable stage from the TODO in the adversarial stage, to better
indicate the current state of the SIAE evaluation. Now that all SIAE
follow the same calendar, simply indicating TODO is sufficient, and
simpler.