employee_record: Truly deterministic order for transfer commands
Previous commit (09c3e912c) was incomplete, we can't rely on
`updated_at` to have a stable order as two objects could be created at
the same time, that what happens in our tests with the `freeze_time()`.
As the database fallback to its row order, creating the object in the
reverse expected order unconditionally trigger the problem.