dashboard: Improve edit_user_email view
Improve user email update:

Configure django-allauth to only allow a user to change their email
address: the user email address is automatically updated when it's verified

Send djanfo-allauth's email but don't log them out.

Before this whange they would receive this email after trying to login
with the new email address.

This is not perfect as the email wording is for signup (but it's the same they used to
receive after trying to login)