Get or create SessionNamespace from its __init__
Make the SessionNamespace less lazy: when it is instantiated, if
initial_data is passed, initialize the SessionNamespace with that data.
Else, load the corresponding SessionNamespace.

Doing so gives a single place to report invalid sessions (missing
session key, session of a different kind), making error handling easier,
rather than needing to call exists() whenever a session is loaded, or
guard all the following calls (get, set, …).

It simplifies the code in most places. The only regression is when the
session key for a job application external transfer does not exist. We
used to redirect to the beginning of the process, we now show a 404.
That’s in line with the rest of the application, and given the hierarchy
for that view, the complexity to maintain that redirection far outweighs
its usefulness.

The use of more precise exceptions instead of raising a 404 when the
session namespace does not exist or is of an unexpected kind allows
users to better handle those cases differently.

The self.expected_session_kind attribute was removed, as __init__
becomes the sole user of that value.