Suspension: change custom rules for the admin
In 2021, we introduced a different behaviour in the Django admin but without testing it:
only end users could change suspension's start date in the past. Admin users were not allowed
to start a suspension before its creation date.
Years went by and admin users forgot this was intended to be this way. They were just blocked.
This commit deletes this rule and adds a test for retrocompatibility purposes.