admin: Allow NIR in searches to contain A or B
The NIR isn’t always decimal, it may contain A or B for Corsica job
seekers.

That may lead the support team to fail to find the user, leading to
incomprehension with users.

The search by truncated NIR wasn’t supported by the rest of the site and
was removed.