Prevent migrations from running concurrently
When multiple instances are serving the site, CleverCloud builds a new
machine for each instance. Part of the build process is to run
CC_MANAGE_PYTHON_TASKS, which includes `migrate`. We don’t want two
concurrent runs of migrations, mostly because data migrations are often
set with `atomic=False`, which means the workers would run them
concurrently.

A session-level advisory lock is acquired for the migrate command to
proceed. It cannot be a transaction-level lock, because some data
migrations run with `atomic=False` and use multiple transactions.
Since the connection will be released at the end of the command, a
session-level advisory lock is guaranteed to be released fairly quickly
after the migrate command completes.

Since the migrations duration cannot be known in advance, the statement
timeout is explicitly unbound. To help troubleshoot a stuck deployment
process, the “Acquiring advisory lock for migrations” message was added.