admin: Make ApprovalAdmin.end_at readonly
This field is tricky. It is maintained via a PostgreSQL trigger, so that
prolongations and suspensions update the `end_at` value. In a typical
support team operation, they would open the Approval admin page (with
the old end_at), then add a prolongation in another tab, comment on the
approval admin page with the Zendesk ticket, and save the approval.
However, the `end_date` has not been refreshed after the prolongation
was saved, so the `end_date` computed by the trigger is overwritten by
the `end_date` submitted by the team member.