Remove a request from can_create_siae_antenna
The perms middleware already fetches organizational data, including
administrator rights. Instead of going back to the database, use the
information set on the request object by the middleware.

This method will soon be called on every page load, as the option to
create an antenna will be moved to the menu, which is rendered on every
page. Saving a request is worth the slightly less intuitive interface of
passing the user administrative status as an argument.

Move the method to a perms module, hoping to consolidate the permissions
into these modules in the future.