Remove a request from can_create_antenna
The perms middleware already fetches organizational data, including
administrator rights. Instead of going back to the database, use the
information set on the request object by the middleware.

Move the method to a perms module, with the hope to consolidate
permissions checks into these modules in the future.

Rename can_create_{siae_,}antenna, as GEIQ can also create antennas, but
are not SIAE.