users: Replace `is_considered_stalled` generated field
Adding a generated field will force a complete table rewrite, which is
not really a good idea to do on a not so small (~1.5M rows) and very
busy table is we don't want to cause errors by blocking everything.
As of now there isn't a way to prevent or circumvent that so we go back
to using classic Django's queryset conditions.