admin: Update manual approval delivery to search approvals
Previously, the PoleEmploiApproval were searched, but we’ll soon drop
that table entirely. The shortcut would be much more useful against the
current approvals anyway, as less than 10 active PoleEmploiApproval
remain.

The markup included extraneous spaces in the parenthesis. Fitting the
anchor tag on a line unblock djlint to properly format this template.