admin: Show the PK for structure display
The PK is often required when filling a form with raw id fields. For
example, a ProlongationRequest requires both the PK of the SIAE and PK
of the prescriber organization. Having the PK displayed on most pages
avoids a lookup (e.g. opening a tab for the structure, only to copy its
PK).