eligibility: Show criteria certification to prescribers
A hiring start date was required in order to compute whether the
certification would be valid at a given date. Prescribers do not know
the hiring start, so they would not know whether their diagnosis was
backed by a state information system.

Employer eligibility diagnosis were made to expire after 92 days
(instead 183), so that the eligibility diagnosis expiry could be used
instead of individual criteria certification period. Which means that we
can simply check whether the criteria was certified and stop checking
the certification_period entirely.