populate_metabase_emplois: Use more appropriate `batch_size`
I used the not modified "organizations" (~4s) and "gps_memberships"
(~3s) as the maximum time limit, all modified `batch_size` make the
chunk be saved in more or less 2s.
I didn't account for RAM but we should be fine as the biggest size
increases are usually for the tables with the less columns or rows.