Do not call certify_criteria from a web worker
That method is already making up to 3 HTTP requests to the API
Particulier. We’re about to add more calls as we certify the RQTH
criteria against France Travail APIs, which itself is 2 to 3 queries.
And the list will grow as we certify more and more criteria.

If any API is slow, the whole site could be taken down as worker spend
their time waiting for API responses.