utils: Don't wrap wet run command in a transaction
I wanted to use this wrapper in fetch_riae_contracts but I wanted
to commit the contracts in bulk (more than 700k lines to sync at each
run).

I don't need it anymore in the management command, but I still think
that we should not have atomic transactions in wet run.