tests: Improve test_certified_criteria_birth_place_not_readonly_if_empty
No need for the job seeker to be born in France if we set their birth
place to None.

Avoid the assertNotContains with markup. It is not paired with a
positive assertion and will become moot when the field markup changes
(which happens fairly often with Django).

Use birth_place_id instead of birth_place, to avoid an extra query from
the related manager.