Create a TEST environment
Allows failing the API particulier requests in tests, and skipping them
during development. We want the failure in tests to easily identify
tests relying on the API particulier (a TypeError when concatenating the
endpoint to the base url, which is None). In development, we don’t want
those errors to show in the web server console.