Show “job seekers in my organization” to 1-member organizations
The special case is not tested and not worth the extra query for all
prescribers (it’s on the nav). If a prescriber is alone in their
organization, the page will not be more useful than “my job seekers”,
but it won’t hurt either.

Drop layout test (navigation bar) for that specific case, since it’s no
longer treated differently.