job_applications: Fix automatic approval delivery
Don't check old approvals to check if we won't to deliver an approval
to a user with a NIR

It's probably a mistake made in commit 588cd37, since the code does not
match the commit message:

JobApplication: automatically create a PASS if no PE Approval is found.
If the user bears:
- a NIR
- or a pole emploi ID
- or is not registered to PE

AND we did not find any PASS or PE Approval that matches one of these
IDs, then consider there is none and deliver the PASS automatically.

In practice, there is actually none, usually, according to the support
team.