itou.utils.triggers: Plan for a context refresh when the manager exit
Once we exit the manager we can't know for sure what is the value in
`itou.context` because we don't know the transaction or savepoint
status, setting `last_data_set` to `None` will force a refresh the next
time we enter the context manager. This is an "lost" extra query if the
transaction was committed but this fix the case where the transaction or
savepoint was rollback and we try to set the same data twice in a row.