eligibility: Certify the RQTH criteria
Update the criteria_certification_badge to use an actual certification
period. Previously, the certification was mostly binary (certified/not
certified), because the API Particulier certifies from the current day,
and has no end date.

The RQTH validity follows calendar spans. To simplify the
implementation, the certification_period does not store the history of
validity periods, but only the active period (when the criteria was
selected in an eligibility diagnosis).
A retry mechanism retries certificating the criteria for job
applications where the certification would be beneficial (not hired
yet).

TODO:
- early PR to create the task before calling it
- maybe reorganize tests/utils/template_tags (to collect all template
  tags and make submodules tests.utils.templatetags.test_xxx)