Employer: Preserve data when reloading hiring confirmation forms
The accept view works as follows:
1. User fills in the fields, POST the data (with HTMX)
2. HTMX reloads the form, with a confirmation modal
3. Confirmation posts the form with an additional "confirmed" field

At step 1, the POST data is passed to the job seeker personal data form,
which updates the form instance [1], that is the view self.job_seeker
[2].
When generating the response for step 2, the data is constructed
from the updated (but not saved) job seeker. The personal data form
reads the birth country from the (unsaved) job seeker profile instance,
and disables the field because the job seeker identity is certified.
On step 3, the form is submitted without disabled fields, loosing the
birth country and triggering a validation error.

[1] https://github.com/django/django/blob/adc80dd81eb70d8e11230b9603b025f0d98e1373/django/forms/models.py#L490-L495
[2] https://github.com/gip-inclusion/les-emplois/blob/13d1d4eba827eec4d17a2560c8d0973325de99ea/itou/www/apply/views/common.py#L49