approvals: only allow to prolong user.latest_approval
With duplicate job seekers it is possible for a job seeker to end up
with two approvals.
And thanks to suspension/prolongation it is also possible to have a
"previous" (start_at sooner) approval still active while the
"is_last_for_user" approval being not "open_to_prolongation" anymore.

Instead, only allow the latest_approval to be prolonged.