GEIQ: Always show the birth country and birth place fields on hire
Avoid conditionally showing this form. Employers should be able to
update the birth_country and birth_place on hire (same as IAE
employers).

We might even want to drop this form to use the same code path as with
IAE employers, but that’s another can of worms and the removal of these
forms is scheduled (and in progress:
https://www.notion.so/gip-inclusion/Refonte-parcours-embauche-1-3-tape-de-r-cup-ration-des-infos-du-candidat-1d55f321b6048046be8ec74bb62ff609).