api_particulier: Handle status code 422
There’s no point in retrying when the API replies it couldn’t find the
person identity record.

404 indicates that the person identity was found (SNGI), but there are
no data provider for that person, or the data provider had no entry for
that person.

422 indicates the person identity was not found, it is meant to
indicate that changing the parameters might help finding that person.

Verbatim from the API particulier support describing this behavior:

> Avant en v2 on avait plusieurs type de 404 :
> - Identité non trouvée (i.e les params fournis ne sont pas suffisant)
> - Pas de caisse de rattachement
> - Pas de dossier dans la caisse

> Maintenant en v3, afin de bien faire la distinction, l’erreur
> d'identité non trouvée est une 422.

> Pourquoi une 422 ?
> Parce que les autres erreurs ont des valeurs métiers concrète
> (L'utilisateur est trouvé, il a ou non une caisse de rattachement et
> il a ou non un dossier).