api_particulier: Quiet down error logging
The generic error logging in place with the API particulier V2 was very
verbose. It was left that way because the error handling flow had many
branches and getting consistent logging was safer.

With the migration to V3, the error handling is much easier to follow,
and we can now respond automatically to most events and log only
interesting events.

Remove the test_service_unavailable, as its only interesting line (the
log) is removed. If the service is unavailable, we will retry later, and
that is tested with
TestCertifyCriteriaApiParticulier::test_retry_task_on_http_error_status_codes.