siaes: fix throttling in SiaeViewSet
There was a specialized class that actually was unused since it was
defined in throttling_classes`but DRF uses `throttle_classes` property.

cf https://www.django-rest-framework.org/api-guide/throttling/

So that particular endpoint was actually following the defaults defined
in the settings, which are in the end pretty okay for our use case as
any authenticated user such as ours gets 120 requests/minute which is
probably fine for our purposes.