job_applications: retrieve GEIQ diagnoses using querysets
In the same way as IAE diagnoses, to be able to display eligibility
badges in prescriptions tables.

One note: the GEIQ eligibility diagnosis badges appear valid when
the considered diagnosis is not expired *and* when it grants an
allowance. Hence, `with_jobseeker_geiq_eligibility_diagnosis` filters
out diagnoses whose criteria or author do not grant the allowance.