eligibility: Introduce async_certify_criterion_with_api_particulier
The eligibility diagnosis isn’t the correct level to handle
certification. If multiple criteria can be certified, a failure for one
criteria would prevent other criteria from being certified.

Instead, deal with each criteria independently.

The task is not called yet, to let it be deployed on the worker first.
It will be called in a future commit.