www.apply: add GEIQ eligibility criteria and badge in prescriptions list
This commit only adds GEIQ criteria in the prescriptions list, which is
straightforward:
as a prescriber or prescribing employer, I only see prescriber GEIQ
diagnoses, which grant the GEIQ allowance de facto.

However we worked on a more complete solution, that would have displayed
GEIQ criteria for received job applications, in list_for_siae. In this
case, a valid diagnosis made by an employer can lead to no allowance. We
developed a logic in the JobApplication querysets to handle this in
previous commits, but due to lack of time, we did not apply it to
list_for_siae.