users.admin: Add a button to reactivate users' account
For now the only allowed case is when the account was deactivated by the
anonymization process for professionnals, because this process is easy
to revert as it only switch `is_active` and empty `email`.
Accounts deactivated by the corresponding button are not supported since
it's more destructive, need more effort to work, and doesn't currently
create a surge of inquiries for the support team.