eligibility: Retry certifying selected administrative criteria
Previously, when a criteria could be certified but wasn’t, we would not
retry the certification.

A retry should not be attempted when the job seeker was hired, because
the hire could evaluated during SIAE evaluations. Since the API answer
is only valid for the day when we make the request, we cannot tell if
person was beneficiary when the hire was made.

Example causes:
- unhandled error from a data provider for the API particulier,
- API particulier unavailable for a day and the retry count was
  exceeded,
- a deploy with changes to certification tasks signature causing task
  invocation to fail (happened a few times in the past)
- a bug such as https://github.com/gip-inclusion/les-emplois/pull/7149
  causing certification tasks not to be spawned,
- Redis becoming unavailable.