eligibility: Record criterion certification request status
The certified boolean field currently indicates whether the API call
succeeded: it is True or False when the person was found (depending on
whether that person was beneficiary), and None when the call failed.

That field will be removed soon: the RQTH certification is based only on
certification_period, which is impossible to express as a boolean.

To correctly convey how to interpret both types of criterion:

- valid for the duration of the eligibility diagnosis (AAH, PI, RSA)
	⇒ the certification_period will be stored as an
	  InclusiveDateRange(certified_at, None), so that the
	  eligibility diagnosis duration sets the limit.

- valid for the certification_period (RQTH)
	⇒ the certification_period will be the actual
	  certification_period as returned by the API.

With the removal of certified=None, it will no longer be possible to
tell whether there was a certification error or the person is not
certified, since both will be represented with
certification_period=None.

The certification error case still needs to be reported to users, that’s
what this new field will be used for.