Eligibility: Remove upper bound from criteria certification_period
Until now, the `certification_period` was not checked by the system,
because the `certified` boolean was sufficient.

With the upcoming RQTH certification, which is based on validity
periods, the certification_period field will become the reference for
certification validity.

The 92 days hack used previously does not work for prescriber diagnosis,
because they are valid 183 days, and we want to show the certified
status for the entire lifetime of the eligibility diagnosis.

Using `None` as the upper bound means the end of the certification is
unknown and will always be considered valid. The system relies on
eligibility diagnosis expiration to hide the certification from users.