Replace SelectedAdministrativeCriteria certified with certification_period
The upcoming RQTH certification will be based on date ranges rather than
being certified for the entire duration of the eligibility diagnosis.

That means we must choose a date to consider when displaying the
certification badge, and check if the chosen date is in the
certification_period.

For API particulier criteria, 3858c068775ed12f39f9a5804dc457f6b8251f58
removed the upper bound so that a certified criteria remains certified
for the entire duration of the eligibility diagnosis.

Now, the certification_period field can be read from to decide on
criteria eligibility. The data model evolved from the previous certified
boolean to certification_period, where:
- `[2025-09-09, NULL)` is roughly equivalent to certified=True
- `()` (empty range) is equivalent to certified=False, the person was
  found be could not be certified.
- `None` is equivalent to certified=None, the person was not found or
  there was an error during the certification.

This commit also removes usage of the certified field, as
certification_period becomes the source of truth.