admin: fix js to scroll down to the current app
Using El.scrollIntoView() with the sticky overflow menu yields a weird
behavior in the last menu entries: it scrolls down not only the menu but
also the entire page, which is not wanted.

Playing with the method's parameters did not bring success.

With this method, no need for the `scrollBy` hack to bring back the
admin header.